Refactor: update imports, improve type consistency, and enhance cDave components with new props and context refinements
