+++ packages/frontend/src/features/cDave/CDaveCard/QQPlot.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useDeepCompareMemo } from 'use-deep-compare';
 import { Loader } from '@mantine/core';
-import EChartWrapper from '@/features/charts/EChartWrapper';
+import EChartWrapper from '../../../components/charts/echarts/EChartWrapper';
 import { EChartsOption } from 'echarts';
 
 const getQuantile = (count: number, quantile: number) =>
